query syntax, method syntax

+++ InformationComputation/DataLayer/Events.cs
@@ -15,12 +15,12 @@ namespace DataLayer
     public abstract class DataAPI
     {
         public abstract UserType GetUser(long id);
-        public abstract float GetPrice(string name);
+        public abstract double GetPrice(string name);
         public abstract bool CheckStorage(string prod_name, int amount);
-        public abstract bool CheckFunds(float price);
+        public abstract bool CheckFunds(double price);
         public abstract void IncreaseStock(string prod_name, int amount);
-        public abstract void RecordProfit(float money);
-        public abstract void AddItem(string name, float price);
+        public abstract void RecordProfit(double money);
+        public abstract void AddItem(string name, double price);
         public abstract void RemoveItem(string name);
         public abstract void AddToStorage(string prod_name, int amount);
         public abstract void RemoveFromStorage(string prod_name);
@@ -38,22 +38,22 @@ namespace DataLayer
 
         public override int GetAmount(string prod_name) => store.GetAmount(prod_name);
         public override UserType GetUser(long id) => users.GetUser(id);
-        public override float GetPrice(string name) => catalog.GetPrice(name);
+        public override double GetPrice(string name) => catalog.GetPrice(name);
         public override bool CheckStorage(string prod_name, int amount) => store.CheckStorage(prod_name, amount);
-        public override bool CheckFunds(float price) => store.CheckFunds(price);
+        public override bool CheckFunds(double price) => store.CheckFunds(price);
         public override void IncreaseStock(string name, int amount) => store.IncreaseStock(name, amount);
-        public override void RecordProfit(float money) => store.RecordProfit(money);
-        public override void AddItem(string name, float price) => catalog.AddItem(name, price);
+        public override void RecordProfit(double money) => store.RecordProfit(money);
+        public override void AddItem(string name, double price) => catalog.AddItem(name, price);
         public override void RemoveItem(string name) => catalog.RemoveItem(name);
         public override void AddToStorage(string name, int amount) => store.AddToStorage(name, amount);
         public override void RemoveFromStorage(string name) => store.RemoveFromStorage(name);
         public override void AddUser(long id, UserType type) => users.AddUser(id, type);
         public override void RemoveUser(long id) => users.RemoveUser(id);
         public override Dictionary<long, UserType> GetAllUsers() => users.UserData;
-        public float GetFunds() => store.GetFunds();
-        public Dictionary<string, (int quantity, float price)> GetInventory()
+        public double GetFunds() => store.GetFunds();
+        public Dictionary<string, (int quantity, double price)> GetInventory()
         {
-            return store.GetInventory(); // Assuming `store` is an instance of Store_state
+            return store.GetInventory();
         }
         public bool ItemExists(string name)
         {
